feat: Rationalize homepage content/form split. Fixes #7
Content should go on pages at the root level or in collections (posts,
teamMembers, etc.). _layouts is just for form and should not include any
real content. Furthermore, the nav bar should be available by default on
every page, so I moved the header to the default layout. I deleted post
and page layouts until we have a need for them, which we currently
don't.